[gpg] Set GPG home directory when reading signature IDs. Fixes JB#44305
GPG will not start unless it finds or can create ~/.gnupg for its conf
and keyrings, and PackageKit runs without a valid home. For every other
key operation it sets the home to temporary directories containing
general or trusted keyrings, but reading a signature ID from a file
shouldn't need one so it's not done right now. Also, zypp shouldn't ever
touch the user's keyrings, as it should operate exclusively with keys
from the RPM db.

This patch sets the keyring to the general keyring home for this call.
It fixes PackageKit, and prevents a permanent ~/.gnupg from being
created.