[libzypp] Do not disable libproxy use. JB#57759
Despite this `-DDISABLE_LIBPROXY=TRUE`, which was added during upgrade
to 17.24.2 with commit 7be7f21, there is also `BuildRequires:
pkgconfig(libproxy-1.0)` left from past. So I guess the option appeared
as new during that upgrade, which was wrongly perceived as that the
feature to use libproxy was new and so it was disabled as unneeded,
missing the fact that it was actually used before.